Refactorizacion

+++ database/migrations/2016_03_19_050846_create_radios_table.php
@@ -21,9 +21,9 @@ class CreateRadiosTable extends Migration
             $table->integer('city_id')->unsigned();
             $table->integer('modulation_id')->unsigned();
             $table->string('name');
-            $table->integer('frequency');
+            $table->float('frequency');
             $table->string('streaming');
-            $table->float('active');
+            $table->integer('active');
             $table->timestamps();
         });
     }
